chore: rename restaurant addresses repositories
